+++ app/Controllers/Designer.php
@@ -251,7 +251,80 @@ class Designer extends BaseController
         }
 
         if ($this->request->getVar('tambah')) {
-            # code...
+            $judul = $this->request->getVar('judul');
+            $kategori = $this->request->getVar('kategori');
+            $harga = $this->request->getVar('harga');
+            $deskripsi = $this->request->getVar('deskripsi');
+            $gambar1 = $this->request->getFile('gambar1');
+            $gambar2 = $this->request->getFile('gambar2');
+            $gambar3 = $this->request->getFile('gambar3');
+            $status = $this->request->getVar('status');
+
+            // MENCARI KATEGORI PRODUK
+            $queryKategoriProduk = $this->builderKategori;
+            $queryKategoriProduk = $queryKategoriProduk->find($kategori);
+            $kategoriProduk = $queryKategoriProduk['kategori'];
+
+            $id = uniqid('prod-' . $kategoriProduk . '-', true);
+
+            // GAMBAR PRODUK
+            $pathUpload = 'asset/produk/' . $kategoriProduk;
+
+            // GAMBAR 1
+            if ($gambar1->getError() == 4) {
+                $gambar1Name = "asset/website/image-default.png";
+            } else {
+                // RENAME FILE
+                $gambar1Name = $gambar1->getRandomName();
+                // MOVE FILE
+                $gambar1->move($pathUpload, $gambar1Name);
+                $gambar1Name = $pathUpload . '/' . $gambar1Name;
+            }
+            // GAMBAR 2
+            if ($gambar2->getError() == 4) {
+                $gambar2Name = "";
+            } else {
+                // RENAME FILE
+                $gambar2Name = $gambar2->getRandomName();
+                // MOVE FILE
+                $gambar2->move($pathUpload, $gambar2Name);
+                $gambar2Name = $pathUpload . '/' . $gambar2Name;
+            }
+            // GAMBAR 3
+            if ($gambar3->getError() == 4) {
+                $gambar3Name = "";
+            } else {
+                // RENAME FILE
+                $gambar3Name = $gambar3->getRandomName();
+                // MOVE FILE
+                $gambar3->move($pathUpload, $gambar3Name);
+                $gambar3Name = $pathUpload . '/' . $gambar3Name;
+            }
+
+            date_default_timezone_set('Asia/Jakarta');
+
+            $dataProduk = [
+                'id' => $id,
+                'judul' => $judul,
+                'harga' => $harga,
+                'deskripsi' => $deskripsi,
+                'gambar1' => $gambar1Name,
+                'gambar2' => $gambar2Name,
+                'gambar3' => $gambar3Name,
+                'status' => $status,
+                'rating' => 0,
+                'terjual' => 0,
+                'created' => date("Y-m-d H:i:s"),
+                'idKategori' => $kategori,
+                'idDesigner' => session()->get('id'),
+            ];
+
+            $queryProduk = $this->builderProduk;
+            if ($queryProduk->insert($dataProduk) === 0) {
+                session()->setFlashdata('add_success', 'Produk Berhasil Ditambahkan!!!');
+            } else {
+                session()->setFlashdata('add_error', 'Produk Gagal Ditambahkan!!!');
+            }
         }
 
         // KATEGORI
@@ -334,7 +407,10 @@ class Designer extends BaseController
             $avatarName = $pathAvatar . $avatarName;
 
             // DELETE FILE LAMA
-            if ($akunLama['avatar'] !== 'asset/designer/akun/avatar-designer.png') {
+            if (
+                $akunLama['avatar'] !== 'asset/designer/akun/avatar-designer.png'
+                && file_exists($akunLama['avatar'])
+            ) {
                 unlink($akunLama['avatar']);
             }
         }
@@ -384,5 +460,139 @@ class Designer extends BaseController
 
             return redirect()->back();
         }
+    }
+
+    public function detailsProduk($id)
+    {
+        // CHECK TIPE AKUN
+        if (session()->get('tipe') !== '2') {
+            return redirect()->to(base_url('/logout'));
+        }
+
+        if ($this->request->getVar('update')) {
+            // PRODUK
+            $queryProduk = $this->builderProduk;
+            $produk = $queryProduk->find($id);
+
+            $judul = $this->request->getVar('judul');
+            $kategori = $this->request->getVar('kategori');
+            $harga = $this->request->getVar('harga');
+            $deskripsi = $this->request->getVar('deskripsi');
+            $gambar1 = $this->request->getFile('gambar1');
+            $gambar2 = $this->request->getFile('gambar2');
+            $gambar3 = $this->request->getFile('gambar3');
+            $status = $this->request->getVar('status');
+            $designerProduk = $this->request->getVar('designer');
+
+            // MENCARI KATEGORI PRODUK
+            $queryKategoriProduk = $this->builderKategori;
+            $queryKategoriProduk = $queryKategoriProduk->find($kategori);
+            $kategoriProduk = $queryKategoriProduk['kategori'];
+
+            // GAMBAR PRODUK
+            $pathUpload = 'asset/produk/' . $kategoriProduk;
+
+            // GAMBAR 1
+            if ($gambar1->getError() == 4) {
+                $gambar1Name = $produk['gambar1'];
+            } else {
+                // RENAME FILE
+                $gambar1Name = $gambar1->getRandomName();
+                // MOVE FILE
+                $gambar1->move($pathUpload, $gambar1Name);
+                $gambar1Name = $pathUpload . '/' . $gambar1Name;
+
+                // DELETE FILE LAMA
+                if (
+                    $produk['gambar1'] !== 'asset/website/image-default.png'
+                    && file_exists($produk['gambar1'])
+                ) {
+                    unlink($produk['gambar1']);
+                }
+            }
+            // GAMBAR 2
+            if ($gambar2->getError() == 4) {
+                $gambar2Name = $produk['gambar2'];
+            } else {
+                // RENAME FILE
+                $gambar2Name = $gambar2->getRandomName();
+                // MOVE FILE
+                $gambar2->move($pathUpload, $gambar2Name);
+                $gambar2Name = $pathUpload . '/' . $gambar2Name;
+
+                // DELETE FILE LAMA
+                if (
+                    $produk['gambar2'] !== 'asset/website/image-default.png'
+                    && file_exists($produk['gambar2'])
+                ) {
+                    unlink($produk['gambar2']);
+                }
+            }
+            // GAMBAR 3
+            if ($gambar3->getError() == 4) {
+                $gambar3Name = $produk['gambar3'];
+            } else {
+                // RENAME FILE
+                $gambar3Name = $gambar3->getRandomName();
+                // MOVE FILE
+                $gambar3->move($pathUpload, $gambar3Name);
+                $gambar3Name = $pathUpload . '/' . $gambar3Name;
+
+                // DELETE FILE LAMA
+                if (
+                    $produk['gambar3'] !== 'asset/website/image-default.png'
+                    && file_exists($produk['gambar3'])
+                ) {
+                    unlink($produk['gambar3']);
+                }
+            }
+
+            $queryUpdateProduk = $this->builderProduk;
+
+            $dataUpdate = [
+                'id' => $id,
+                'judul' => $judul,
+                'harga' => $harga,
+                'deskripsi' => $deskripsi,
+                'gambar1' => $gambar1Name,
+                'gambar2' => $gambar2Name,
+                'gambar3' => $gambar3Name,
+                'status' => $status,
+                'rating' => $produk['rating'],
+                'terjual' => $produk['terjual'],
+                'idKategori' => $kategori,
+                'idDesigner' => $designerProduk,
+            ];
+
+            if ($queryUpdateProduk->save($dataUpdate)) {
+                session()->setFlashdata('update_success', 'Produk Berhasil Diubah!!!');
+            } else {
+                session()->setFlashdata('update_error', 'Produk Gagal Diubah!!!');
+            }
+        }
+
+        // KATEGORI
+        $queryKategori = $this->builderKategori;
+        $kategori = $queryKategori->findAll();
+
+        // PRODUK
+        $queryProduk = $this->builderProduk;
+        $produk = $queryProduk->find($id);
+
+        // DESIGNER
+        $queryDesigner = $this->builderDesigner;
+        $designer = $queryDesigner->findAll();
+
+        $akun = $this->builderAkun->find(session()->get('id'));
+        $data = [
+            'title' => 'Detail Produk | Rajasa Finising',
+            'akun' => $akun,
+            'segment2' => $this->uri->getSegment(2),
+            'kategori' => $kategori,
+            'designer' => $designer,
+            'produk' => $produk,
+        ];
+
+        return view('designer/data-produk/detail', $data);
     }
 }
